[RISCV] add support for scalable vector in gather/scatter instruction costmodel
The origin gather/scatter instruction costmodel only support fix-lenght vector. We merge it with the costmodel that supprot scalable vector from llvm-epi.
The scalable gather/scatter cost is now set as 1 because we can use only one instruction to finish it, but the actual cost may be higher. And to deal with the shuffle vector instruction use in scalable gather/scatter, we add support for scalable vector in shuffle vector costmodel.